Killmenow

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,6 +1,7 @@
 # create a process named killmenow 
 
-exec { 'pkill killmenow' :
-    path    => '/bin/',
-    command => 'pkill killmenow',
-    }
+exec { 'killmenow':
+  command     => 'pkill killmenow',
+  path        => '/usr/local/bin:/usr/bin:/bin',
+  onlyif      => 'pgrep killmenow',
+}
